+++ src/helpers/conversions.ts
@@ -3,7 +3,7 @@ import _ from "lodash";
 import { LANGUAGES } from "../";
 import UnitsEn from "../../nlp/units_EN.json";
 import UnitsFr from "../../nlp/units_FR.json";
-import { globalUnit } from "./models_FR";
+import { globalUnit } from "./keys";
 
 export default class ConversionsUtils {
   /**
@@ -46,22 +46,38 @@ export default class ConversionsUtils {
       out = (this.getUnits(lang).wordNumbers as any)[val];
     }
 
-    // Summing up moyen amounts (2-3 oeufs, 4-5 grammes, ...)
-    const moyenRec = /(\d)?\-(\d)/gim;
-    const matchesMoyen = moyenRec.exec(String(val));
+    //  Normalizing and multiply word and number amounts (3 demi oeufs, 3 demi poulet, ...)
+    const mutlipleRec = /(\d+) (\w*)/gim;
+    const matchesMultiple = mutlipleRec.exec(String(val));
 
-    if (matchesMoyen) {
-      if (matchesMoyen[1] && matchesMoyen[2]) {
-        out = (Number(matchesMoyen[1]) + Number(matchesMoyen[2])) / 2;
+    if (matchesMultiple) {
+      if (
+        matchesMultiple[1] &&
+        matchesMultiple[2] &&
+        this.getUnits(lang).wordNumbers.hasOwnProperty(matchesMultiple[2])
+      ) {
+        out =
+          Number(matchesMultiple[1]) *
+          (this.getUnits(lang).wordNumbers as any)[matchesMultiple[2]];
+      }
+    }
+
+    // Summing up average amounts (2-3 oeufs, 4-5 grammes, ...)
+    const averageRec = /(\d)?\-(\d)/gim;
+    const matchesAverage = averageRec.exec(String(val));
+
+    if (matchesAverage) {
+      if (matchesAverage[1] && matchesAverage[2]) {
+        out = (Number(matchesAverage[1]) + Number(matchesAverage[2])) / 2;
       }
     }
 
     // Summing up mixed amounts (integer and fractions)
-    const fractionRe = /(\d)?(.*)?(\d)\/(\d)/gim;
+    const fractionRe = /(\d+)?(.*)?(\d)\/(\d)/gim;
     const matches = fractionRe.exec(String(val));
     if (matches) {
       if (matches[1] && matches[3] && matches[4]) {
-        out = Number(matches[1]) + Number(matches[3]) / Number(matches[4]);
+        out = (Number(matches[1]) * Number(matches[3])) / Number(matches[4]);
       } else {
         out = Number(matches[3]) / Number(matches[4]);
       }