+++ CHANGELOG.md
@@ -1,3 +1,15 @@
 ## v1.0.0
 
 - Release
+
+## v1.3.6
+
+- Add new support keys for units detection
+
+## v1.3.7
+
+- Change return key unit to EN for all language (EN, FR)
+
+- Add Support of word and fraction number detection (5 1/2 liter of milk)
+
+- Add Support word and number detection detection (5 quarter of oragne)